% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms__aa_properties.R
\name{tardbpdms__aa_properties}
\alias{tardbpdms__aa_properties}
\title{tardbpdms__aa_properties}
\usage{
tardbpdms__aa_properties(aa_properties_file, selected_identifiers = NULL)
}
\arguments{
\item{aa_properties_file}{path to amino acid properties file (required)}

\item{selected_identifiers}{path to file with selected subset of identifiers}
}
\value{
results from PCA
}
\description{
Create results folder for analysis script plots and saved objects.
}
