% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms__gg_color_hue.R
\name{tardbpdms__gg_color_hue}
\alias{tardbpdms__gg_color_hue}
\title{tardbpdms__gg_color_hue}
\usage{
tardbpdms__gg_color_hue(n)
}
\arguments{
\item{n}{an integer number of colours required (required)}
}
\value{
a vector of colours
}
\description{
ggplot2-like colour scale in HCL space.
}
