% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms__ggpairs_binhex.R
\name{tardbpdms__ggpairs_binhex}
\alias{tardbpdms__ggpairs_binhex}
\title{tardbpdms__ggpairs_binhex}
\usage{
tardbpdms__ggpairs_binhex(input_dt, output_file, input_dt_upper = NULL,
  width = 12, height = 12, bins = 50, xlab = "x", ylab = "y",
  title = "", label_size = 4, plot_colours = c("lightgrey", "black"),
  colour_limits = NA)
}
\arguments{
\item{input_dt}{input data table (required)}

\item{output_file}{plot output path (required)}

\item{width}{plot width (default: 12)}

\item{height}{plot height (default: 12)}

\item{bins}{number of hexagon bins (default: 50)}

\item{xlab}{the title of the x-axis}

\item{ylab}{the title of the y-axis}

\item{title}{the plot title}

\item{label_size}{the size of the label text}

\item{plot_colours}{the lower and upper gradient limit colours}

\item{colour_limits}{the colour gradient limits}
}
\value{
Nothing
}
\description{
GGpairs plot for all variables in input data table with correlation in upper triangle and 2d binned hexagons in lower triangle.
}
