% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms__glm_prenorm.R
\name{tardbpdms__glm_prenorm}
\alias{tardbpdms__glm_prenorm}
\title{tardbpdms__glm_prenorm}
\usage{
tardbpdms__glm_prenorm(x)
}
\arguments{
\item{x}{data.frame to be normalised (required)}
}
\value{
A normalised data.frame
}
\description{
Normalise data.frame columns.
}
