% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms__guenther_structure_propensity_plot.R
\name{tardbpdms__guenther_structure_propensity_plot}
\alias{tardbpdms__guenther_structure_propensity_plot}
\title{tardbpdms__guenther_structure_propensity_plot}
\usage{
tardbpdms__guenther_structure_propensity_plot(kernel_scores, outpath,
  colour_scheme, pdb_names, score_type, pdb_list, Nsamples = 10000,
  position_offset = 1, plot_ylim = c(0, 4.5), width = 10,
  height = 5)
}
\arguments{
\item{kernel_scores}{(required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{pdb_names}{character vector of PDB ids}

\item{score_type}{name of score}

\item{pdb_list}{list of lists}

\item{Nsamples}{number of randomisations for score significance specificity test}

\item{position_offset}{offset position}

\item{plot_ylim}{y-axis limits}

\item{width}{plot width (default:6)}

\item{height}{plot height (default:6)}
}
\value{
Nothing
}
\description{
Plot kernel structure propensity results.
}
