% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms__load_toxicity.R
\name{tardbpdms__load_toxicity}
\alias{tardbpdms__load_toxicity}
\title{tardbpdms__load_toxicity}
\usage{
tardbpdms__load_toxicity(dms_dir, object = F, read_filter = T)
}
\arguments{
\item{dms_dir}{path to variant toxicity directory (required)}

\item{object}{load from R data object (default:F)}

\item{read_filter}{filter reads based on is.reads0 (default:T)}
}
\value{
a variant toxicity data.table
}
\description{
Load variant toxicity estimates.
}
