% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms__plot_toxicity_distributions.R
\name{tardbpdms__plot_toxicity_distributions}
\alias{tardbpdms__plot_toxicity_distributions}
\title{tardbpdms__plot_toxicity_distributions}
\usage{
tardbpdms__plot_toxicity_distributions(input_dt, output_file,
  boundary_list = NULL, boundary_colour_list = NULL, xlim = NULL,
  colour_scheme = NULL, width = 6, height = 13,
  facet_by_mut_type = T)
}
\arguments{
\item{input_dt}{a data.table with columns: STOP, toxicity_cond, region, mut_type (required)}

\item{output_file}{path to output plot (required)}

\item{boundary_list}{a list of toxicity values to highlight with vertical lines on plot}

\item{boundary_colour_list}{a list of colours corresponding the boundary_list vertical lines}

\item{xlim}{x axis limits}

\item{colour_scheme}{colour scheme list}

\item{width}{plot width}

\item{height}{plot height}

\item{facet_by_mut_type}{facet by mut_type (default:T)}
}
\value{
Nothing
}
\description{
Plot toxicity distributions (and optional vertical lines highlighting specific toxicity values).
}
