% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms__plot_toxicity_vs_position.R
\name{tardbpdms__plot_toxicity_vs_position}
\alias{tardbpdms__plot_toxicity_vs_position}
\title{tardbpdms__plot_toxicity_vs_position}
\usage{
tardbpdms__plot_toxicity_vs_position(input_dt, output_file, ylab,
  use_sigma = F, span = 0.75, label_offset = 0.05, plot_rect = F,
  plot_hotspot = F)
}
\arguments{
\item{input_dt}{a data.table with columns: toxicity, sigma, region, Pos (required)}

\item{output_file}{path to output plot (required)}

\item{ylab}{y-axis title (required)}

\item{use_sigma}{whether to use error in 'sigma' column (default:F i.e. error estimated from toxicity distribution)}

\item{span}{loess smoother span}

\item{label_offset}{annotation label offset}

\item{plot_rect}{whether to plot}

\item{plot_hotspot}{whether to plot hotspot}
}
\value{
Nothing
}
\description{
Plot toxicity versus position.
}
