% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms__scale_toxicity.R
\name{tardbpdms__scale_toxicity}
\alias{tardbpdms__scale_toxicity}
\title{tardbpdms__scale_toxicity}
\usage{
tardbpdms__scale_toxicity(toxicity_path, outpath)
}
\arguments{
\item{toxicity_path}{folder path to toxicity estimates (required)}

\item{outpath}{output path for plots and saved objects (required)}
}
\value{
Nothing
}
\description{
Scale toxicity estimates between replicates.
}
