% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms__single_mutant_heatmap.R
\name{tardbpdms__single_mutant_heatmap}
\alias{tardbpdms__single_mutant_heatmap}
\title{tardbpdms__single_mutant_heatmap}
\usage{
tardbpdms__single_mutant_heatmap(input_df, variable_name = "fitness",
  mut_dict = NULL, text_size = 2.5, output_file, colour_low = "blue",
  colour_high = "red", width = 12, height = 4,
  x_annotation = "position", xaxis_angle = 90, na_colour = "grey50",
  na_text = "", cluster = "none")
}
\arguments{
\item{input_df}{matrix of heatmap values (required)}

\item{variable_name}{name of variable to use for heatmap cells (defaut:"fitness")}

\item{mut_dict}{list of mutant codes where name is used to label the corresponding cells}

\item{text_size}{text size (default:2.5)}

\item{output_file}{path to output file (required)}

\item{colour_low}{colour scale lower limit colour -- passed to scale_colour_gradient2 (default:"blue")}

\item{colour_high}{colour scale upper limit colour -- passed to scale_colour_gradient2 (default:"red")}

\item{width}{plot width in "units" (default:12)}

\item{height}{plot height in "units" (default:4)}

\item{x_annotation}{x-axis title (default:"position")}

\item{xaxis_angle}{rotation angle for x tick labels -- passed to element_text (default:90)}

\item{na_colour}{colour to use for NA values (default:"grey50")}

\item{na_text}{cell text annotation to use for NA values (default:empty string)}

\item{cluster}{heirarchically cluster (default:none)}
}
\value{
Nothing
}
\description{
ggplot tile heatmap wrapper.
}
