% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms__toxicity_lm_summary.R
\name{tardbpdms__toxicity_lm_summary}
\alias{tardbpdms__toxicity_lm_summary}
\title{tardbpdms__toxicity_lm_summary}
\usage{
tardbpdms__toxicity_lm_summary(input_df, output_file, return_results = F,
  colour_scheme = NULL, subset = NULL, width = 6, height = 6)
}
\arguments{
\item{input_df}{data.frame with toxicity and predictor variables (required)}

\item{output_file}{path for output plot file (required)}

\item{return_results}{whether or not to return rsq_list (default:F)}

\item{colour_scheme}{colour scheme list}

\item{subset}{data subset (default:none)}

\item{width}{plot width (default:6)}

\item{height}{plot height (default:6)}
}
\value{
Nothing
}
\description{
Dot plots showing explained variance of models to predict variant toxicity.
}
