% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms_aa_properties_mutant_effects.R
\name{tardbpdms_aa_properties_mutant_effects}
\alias{tardbpdms_aa_properties_mutant_effects}
\title{tardbpdms_aa_properties_mutant_effects}
\usage{
tardbpdms_aa_properties_mutant_effects(toxicity_dt, outpath, aaprop_file,
  aaprop_file_selected, colour_scheme, execute = TRUE)
}
\arguments{
\item{toxicity_dt}{data.table with single and double mutant codes (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{aaprop_file}{path to amino acid properties file (required)}

\item{aaprop_file_selected}{path to file with selected subset of identifiers}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
A data.table with single and double mutant effects from AA PCA
}
\description{
Calculate single and double mutant effects from AA PCA (using single mutants).
}
