% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms_combine_toxicity.R
\name{tardbpdms_combine_toxicity}
\alias{tardbpdms_combine_toxicity}
\title{tardbpdms_combine_toxicity}
\usage{
tardbpdms_combine_toxicity(toxicity_list, growth_rate_file, outpath,
  colour_scheme, execute = TRUE)
}
\arguments{
\item{toxicity_list}{named list of folder paths with toxicity estimates (required)}

\item{growth_rate_file}{path to growth rate file (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
A data.table with normalised toxicity data
}
\description{
Combine toxicity estimates from two different DMS experiments (regions).
}
