% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms_dimsumcounts_to_fitness.R
\name{tardbpdms_dimsumcounts_to_fitness}
\alias{tardbpdms_dimsumcounts_to_fitness}
\title{tardbpdms_dimsumcounts_to_fitness}
\usage{
tardbpdms_dimsumcounts_to_fitness(counts_path, outpath, generations_path,
  all_reps, min_mean_input_read_count = 10,
  min_input_read_count_doubles = 50, lam_d = 0.025, numCores = 1,
  execute = TRUE)
}
\arguments{
\item{counts_path}{path to DiMSum output (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{generations_path}{path to generations data (required)}

\item{all_reps}{list of replicates to retain (required)}

\item{min_mean_input_read_count}{minimum mean input read count for high confidence variants (default:10)}

\item{min_input_read_count_doubles}{minimum input read count for doubles used to derive prior for Bayesian doubles correction (default:50)}

\item{lam_d}{Poisson distribution for score likelihood (default:0.025)}

\item{numCores}{Number of available CPU cores (default:1)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Estimate fitness of singles, doubles and silent AA substitution variants from DiMSum formatted output counts.
}
