% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms_epistasis_analysis.R
\name{tardbpdms_epistasis_analysis}
\alias{tardbpdms_epistasis_analysis}
\title{tardbpdms_epistasis_analysis}
\usage{
tardbpdms_epistasis_analysis(toxicity_path, miscpath, DMS2structure_path,
  numCores = 1, execute = TRUE)
}
\arguments{
\item{toxicity_path}{path to R data object with single and double mutant toxicity values (required)}

\item{miscpath}{path to misc scripts and data directory (required)}

\item{DMS2structure_path}{Path to DMS2structure repository (required)}

\item{numCores}{Number of available CPU cores (default:1)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Epistasis analysis.
}
