% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms_num_introduced_aa_violins.R
\name{tardbpdms_num_introduced_aa_violins}
\alias{tardbpdms_num_introduced_aa_violins}
\title{tardbpdms_num_introduced_aa_violins}
\usage{
tardbpdms_num_introduced_aa_violins(toxicity_dt, outpath, colour_scheme,
  execute = TRUE)
}
\arguments{
\item{toxicity_dt}{data.table with single and double mutant toxicity values and their mutant effects from AA PCA (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Violin plots showing (residual) toxicity versus number of introduced AAs of various properties.
}
