% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms_PWI_heatmaps.R
\name{tardbpdms_PWI_heatmaps}
\alias{tardbpdms_PWI_heatmaps}
\title{tardbpdms_PWI_heatmaps}
\usage{
tardbpdms_PWI_heatmaps(toxicity_dt, PWI_dir_list, outpath, colour_scheme,
  execute = TRUE)
}
\arguments{
\item{toxicity_dt}{data.table with single and double mutant toxicity values (required)}

\item{PWI_dir_list}{list of PWI directories (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Hydrophobicity of WT and toxicity hotspot line plot.
}
