% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms_quality_control.R
\name{tardbpdms_quality_control}
\alias{tardbpdms_quality_control}
\title{tardbpdms_quality_control}
\usage{
tardbpdms_quality_control(toxicity_list, outpath, colour_scheme,
  execute = TRUE)
}
\arguments{
\item{toxicity_list}{named list of folder paths with toxicity estimates (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Quality control plots including scatterplots comparing toxicity estimates between replicates.
}
