% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms_single_mutant_heatmaps.R
\name{tardbpdms_single_mutant_heatmaps}
\alias{tardbpdms_single_mutant_heatmaps}
\title{tardbpdms_single_mutant_heatmaps}
\usage{
tardbpdms_single_mutant_heatmaps(toxicity_dt, disease_mut_file, outpath,
  colour_scheme, execute = TRUE)
}
\arguments{
\item{toxicity_dt}{data.table with single mutant toxicity values (required)}

\item{disease_mut_file}{table of human disease mutations and classifications (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Combine toxicity estimates from two different DMS experiments (regions).
}
