% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms__plot_kernel_structure_propensity.R,
%   R/tardbpdms_toxicity_model_summary.R
\name{tardbpdms_toxicity_model_summary}
\alias{tardbpdms_toxicity_model_summary}
\title{tardbpdms_toxicity_model_summary}
\usage{
tardbpdms_toxicity_model_summary(toxicity_dt, outpath, colour_scheme,
  execute = TRUE)

tardbpdms_toxicity_model_summary(toxicity_dt, outpath, colour_scheme,
  execute = TRUE)
}
\arguments{
\item{toxicity_dt}{data.table with single and double mutant toxicity values and their mutant effects from AA PCA (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}

\item{toxicity_dt}{data.table with single and double mutant toxicity values and their mutant effects from AA PCA (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing

Nothing
}
\description{
Dot plots showing explained variance of models to predict variant toxicity.

Dot plots showing explained variance of models to predict variant toxicity.
}
