% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tardbpdms_secondary_structure_predictions.R
\name{tardbpdms_secondary_structure_predictions}
\alias{tardbpdms_secondary_structure_predictions}
\title{tardbpdms_secondary_structure_predictions}
\usage{
tardbpdms_secondary_structure_predictions(toxicity_dt, ss_result_file,
  outpath, colour_scheme, execute = TRUE)
}
\arguments{
\item{toxicity_dt}{data.table with single and double mutant toxicity values (required)}

\item{ss_result_file}{RData object with secondary structure 'result_list' data.table (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Secondary structure predictions.
}
